pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gated-agent-protocol"
version = "0.1.0"
description = "The Sovereign Gated Agent Protocol Engine."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0"
]

[project.scripts]
gap = "gap.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/gap"]

[tool.hatch.build]
include = [
  "src/gap/protocols/**/*"
]
